Reorganise solution
