+++ .gitignore
@@ -1,1 +1,3 @@
 .vs/
+obj/
+bin/
