+++ src/MediaReplacement.TestSite/Umbraco.Community.MediaReplacement.TestSite.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
@@ -18,6 +18,10 @@
 
   <ItemGroup>
     <Folder Include="wwwroot\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Umbraco.Community.MediaReplacement.Services\Umbraco.Community.MediaReplacement.Services.csproj" />
   </ItemGroup>
 
   <PropertyGroup>
