+++ src/Umbraco.Community.MediaReplacement.Services/Controllers/MediaFoldersController.cs
@@ -1,5 +1,6 @@
 ﻿using Asp.Versioning;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Umbraco.Cms.Api.Common.ViewModels.Pagination;
 using Umbraco.Cms.Api.Management.Controllers.Media.Tree;
@@ -9,15 +10,13 @@ using Umbraco.Cms.Api.Management.Services.Entities;
 using Umbraco.Cms.Api.Management.ViewModels.Tree;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Cache;
-using Umbraco.Cms.Core.Composing;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.Entities;
-using Umbraco.Cms.Core.Persistence.Querying;
 using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Web.Common.Authorization;
 
-namespace MediaSection.Controllers
+namespace Umbraco.Community.MediaReplacement.Services.Controllers
 {
 
 	[ApiController]
@@ -74,56 +73,6 @@ namespace MediaSection.Controllers
 			MediaTreeItemResponseModel[] treeItemViewModels = MapTreeItemViewModels(parentId, children);
 			PagedViewModel<MediaTreeItemResponseModel> result = PagedViewModel(treeItemViewModels, totalItems);
 			return Task.FromResult<ActionResult<PagedViewModel<MediaTreeItemResponseModel>>>(Ok(result));
-		}
-	}
-
-	public interface IFilteredEntityService
-	{
-		IEnumerable<IEntitySlim> GetPagedChildren(
-			 Guid? parentKey,
-			 UmbracoObjectTypes childObjectType,
-			 int skip,
-			 int take,
-			 out long totalRecords,
-			 IQuery<IUmbracoEntity>? filter = null,
-			 Ordering? ordering = null);
-	}
-
-	public class MediaFoldersOnlyFilteredEntityService : IFilteredEntityService
-	{
-		private readonly IEntityService _entityService;
-		
-		public MediaFoldersOnlyFilteredEntityService(IEntityService entityService)
-		{
-			_entityService = entityService;
-		}
-		
-		public IEnumerable<IEntitySlim> GetPagedChildren(
-			Guid? parentKey,
-			UmbracoObjectTypes childObjectType,
-			int skip,
-			int take,
-			out long totalRecords,
-			IQuery<IUmbracoEntity>? filter = null,
-			Ordering? ordering = null)
-		{
-			return
-				_entityService
-					.GetPagedChildren(parentKey, childObjectType, skip, take, out totalRecords, filter, ordering)
-					.Where(FoldersOnlyFilter);
-		}
-
-		private bool FoldersOnlyFilter(IEntitySlim e) => e is IMediaEntitySlim && ((IMediaEntitySlim)e).ContentTypeKey == new Guid("f38bd2d7-65d0-48e6-95dc-87ce06ec2d3d");// TODO : Don't hardcode
-
-	}
-
-
-	public class MediaApiComposer : IComposer // TODO : Split
-	{
-		public void Compose(IUmbracoBuilder builder)
-		{
-
-			builder.Services.AddTransient<IFilteredEntityService, MediaFoldersOnlyFilteredEntityService>();
 		}
 	}
 }
