+++ src/Umbraco.Community.MediaReplacement.Services/Umbraco.Community.MediaReplacement.Services.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Umbraco.Cms" Version="16.1.0" />
+	</ItemGroup>
+
+</Project>
